LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/function-shadow-container-expected.txt


PASS Can query named container in shadow
FAIL ::part() can see inner named containers assert_equals: expected "5px" but got "10px"
FAIL ::slotted() can see inner named containers assert_equals: expected "5px 5px" but got "10px 10px"
